// src/lib/types.ts
export type ConrodItem = {
  id: number
  serialNumber: string
  conrodName: string
  conrodVariant: string
  conrodSize: string
  smallEndDiameter: number
  bigEndDiameter: number
  centerDistance: number
  pinName: string
  pinSize: string
  ballBearingName: string
  ballBearingVariant: string
  ballBearingSize: string
  amount?: number
  createdAt?: string
}